Aggiunta chiamata put ad AreaController, possibile modificare un area

+++ eventManagement/src/main/java/it/unibo/cas/eventmanagement/repositories/AreaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AreaRepository extends JpaRepository<Area, Integer> {
+public interface AreaRepository extends JpaRepository<Area, Long> {
 
     Area getAreaByName(String name);
